UNFINISHED - Reverting changes, back to Jack's Master
NB: There is more to line 44 than just "uglify the .js". For a start the var statement near the beginning is structured differently (not just var $j=jQuery.noConflict(),) , and the end also seems hybrid, with the content from the .js plus some more... I have to learn a bit about all this.